src/api/requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
aiofiles==24.1.0

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

# GraphRAG + related
nano-graphrag @ git+https://github.com/ECE496-Team-DASH/nano-graphrag.git@main
nano-vectordb>=0.0.4.3
tiktoken==0.7.0
networkx==3.3
openai>=1.51.2
google-genai>=0.4.0
tenacity>=8.5.0

# PDF + arXiv
PyMuPDF==1.24.10
arxiv==2.1.3
httpx==0.27.2

python-dotenv==1.0.1
portalocker==2.10.1
